Added updated scraper.py and token_extractor.py
scraper.py is now updated with better regular expressions. scraper still needs functionality for tokenizing sentences and storing each sentence into files. scraper also needs to be able to work with multiple urls, not just one. 

token_extractor is a function that extracts important terms from the parsed web pages using an importance measure such as term frequency. This function will build a vocabulary of unique terms and create a dictionary of unique terms where the key is the token and the value is the count across all documents.  I still need to implement functionality to print the top 25-40 terms. At the moment this function only returns the frequency distribution and the dictionary vocabulary.