updated token_extractor.py
Added functionality to sum up multiple dictionaries where each dictionary corresponds to a clean text file. This function can now get top 40 terms from that cumulative dictionary.

Also, added small code to remove all punctuation, as they are needed according to the project description. This function is mainly used for debugging purposes if you want to play with it. The official function will be implemented in the webCrawler.py